requirements.txt
Flask==3.0.3
pdf2docx==0.5.8
python-docx==1.1.2
Werkzeug==3.0.3
pypdf==4.2.0
Pillow==10.4.0
pdf2image==1.17.0
ocrmypdf==16.6.1